# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AtlasGuard API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface WeightConfig {
  affordability: number;
  climate: number;
  safety: number;
  amenities: number;
  familyFit: number;
  mobility: number;
  opportunity: number;
}

export interface IncomeConfig {
  current: number;
  target: number;
}

export interface HousingConfig {
  current: number;
  target: number;
}

export interface ExpensesConfig {
  current: number;
  target: number;
}

export type EvaluateRequestPriorityMode =
  (typeof EvaluateRequestPriorityMode)[keyof typeof EvaluateRequestPriorityMode];

export const EvaluateRequestPriorityMode = {
  finance: "finance",
  balanced: "balanced",
  lifestyle: "lifestyle",
} as const;

export type EvaluateRequestClimatePreference =
  (typeof EvaluateRequestClimatePreference)[keyof typeof EvaluateRequestClimatePreference];

export const EvaluateRequestClimatePreference = {
  warm: "warm",
  cold: "cold",
} as const;

export interface EvaluateRequest {
  /** @nullable */
  email?: string | null;
  currentCity: string;
  targetCity: string;
  priorityMode: EvaluateRequestPriorityMode;
  climatePreference: EvaluateRequestClimatePreference;
  weights: WeightConfig;
  income: IncomeConfig;
  housing: HousingConfig;
  expenses: ExpensesConfig;
  retainedIncomeAdjustment: number;
}

export type EvaluateResponseVerdict =
  (typeof EvaluateResponseVerdict)[keyof typeof EvaluateResponseVerdict];

export const EvaluateResponseVerdict = {
  strong_fit: "strong_fit",
  qualified_yes: "qualified_yes",
  caution: "caution",
  weak_fit: "weak_fit",
} as const;

export type EvaluateResponseResultMode =
  (typeof EvaluateResponseResultMode)[keyof typeof EvaluateResponseResultMode];

export const EvaluateResponseResultMode = {
  explainer: "explainer",
  repaired_explainer: "repaired_explainer",
  deterministic_fallback: "deterministic_fallback",
} as const;

export type EvaluateResponseStructuredExplanation = { [key: string]: unknown };

export type EvaluateResponseRenderedOutput = { [key: string]: unknown };

export interface EvaluateResponse {
  scenarioId: number;
  lifestyleFit: number;
  financialFit: number;
  moveScore: number;
  verdictBand: string;
  verdict: EvaluateResponseVerdict;
  resultMode: EvaluateResponseResultMode;
  structuredExplanation?: EvaluateResponseStructuredExplanation;
  renderedOutput?: EvaluateResponseRenderedOutput;
}

export interface ScenarioSummary {
  id: number;
  currentCity: string;
  targetCity: string;
  createdAt: string;
}

export interface ScenarioDetail {
  id: number;
  currentCity: string;
  targetCity: string;
  /** @nullable */
  email?: string | null;
  createdAt: string;
  result?: EvaluateResponse;
}

export interface Metro {
  id: number;
  city: string;
  state: string;
  slug: string;
  affordabilityScore: number;
  climateScore: number;
  climateTendency: string;
  safetyScore: number;
  amenitiesScore: number;
  familyFitScore: number;
  mobilityScore: number;
  opportunityScore: number;
  incomeTaxRegime: string;
  benchmarkVersion: string;
  staleRiskFlag: boolean;
}