# lib/api-zod/src/generated/types/evaluateRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AtlasGuard API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EvaluateRequestClimatePreference } from "./evaluateRequestClimatePreference";
import type { EvaluateRequestPriorityMode } from "./evaluateRequestPriorityMode";
import type { ExpensesConfig } from "./expensesConfig";
import type { HousingConfig } from "./housingConfig";
import type { IncomeConfig } from "./incomeConfig";
import type { WeightConfig } from "./weightConfig";

export interface EvaluateRequest {
  /** @nullable */
  email?: string | null;
  currentCity: string;
  targetCity: string;
  priorityMode: EvaluateRequestPriorityMode;
  climatePreference: EvaluateRequestClimatePreference;
  weights: WeightConfig;
  income: IncomeConfig;
  housing: HousingConfig;
  expenses: ExpensesConfig;
  retainedIncomeAdjustment: number;
}