# lib/api-zod/src/generated/types/evaluateResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AtlasGuard API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EvaluateResponseRenderedOutput } from "./evaluateResponseRenderedOutput";
import type { EvaluateResponseResultMode } from "./evaluateResponseResultMode";
import type { EvaluateResponseStructuredExplanation } from "./evaluateResponseStructuredExplanation";
import type { EvaluateResponseVerdict } from "./evaluateResponseVerdict";

export interface EvaluateResponse {
  scenarioId: number;
  lifestyleFit: number;
  financialFit: number;
  moveScore: number;
  verdictBand: string;
  verdict: EvaluateResponseVerdict;
  resultMode: EvaluateResponseResultMode;
  structuredExplanation?: EvaluateResponseStructuredExplanation;
  renderedOutput?: EvaluateResponseRenderedOutput;
}