# lib/api-zod/src/generated/types/expensesConfig.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AtlasGuard API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ExpensesConfig {
  current: number;
  target: number;
}