# lib/api-zod/src/generated/types/scenarioDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AtlasGuard API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EvaluateResponse } from "./evaluateResponse";

export interface ScenarioDetail {
  id: number;
  currentCity: string;
  targetCity: string;
  /** @nullable */
  email?: string | null;
  createdAt: Date;
  result?: EvaluateResponse;
}