# lib/api-zod/src/generated/types/scenarioSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AtlasGuard API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScenarioSummary {
  id: number;
  currentCity: string;
  targetCity: string;
  createdAt: Date;
}