# lib/api-zod/src/generated/types/weightConfig.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AtlasGuard API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WeightConfig {
  affordability: number;
  climate: number;
  safety: number;
  amenities: number;
  familyFit: number;
  mobility: number;
  opportunity: number;
}